prime numbers solution

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,9 +8,9 @@
  */
 int main(void)
 {
-	long long int n;
-	long long int max;
-	long long int i;
+	long int n;
+	long int max;
+	long int i;
 
 	n = 612852475143;
 	max = -1;
